requirements.txt
google-adk>=0.3.0
google-generativeai>=0.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
PyPDF2>=3.0.0
reportlab>=4.0.0
pypdf>=3.17.0
feedparser>=6.0.10
tweepy>=4.14.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0

